修复下载错误不报错的问题

+++ APKInstaller/APKInstaller/Helpers/ProgressHelper.cs
@@ -74,7 +74,7 @@ namespace APKInstaller.Helpers
 
         private static void SetProgressState(ProgressState state)
         {
-            _taskbarList?.SetProgressState(GetHandle(), (TBPFLAG)Enum.Parse(typeof(TBPFLAG), state.ToString()));
+            _taskbarList?.SetProgressState(GetHandle(), (TBPFLAG)Enum.Parse(typeof(TBPFLAG), ((int)state).ToString()));
         }
 
         private static void SetProgressValue(int current, int max)
